% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{AppData}
\alias{AppData}
\title{Get an application data directory}
\usage{
AppData(pkgname = "SeuratData", author = pkgname)
}
\arguments{
\item{pkgname}{Name of package}

\item{author}{Author name for application}
}
\value{
A character vector with path to the application data dir
}
\description{
Get an application data directory
}
\seealso{
\code{\link[rappdirs]{user_data_dir}}
}
\keyword{internal}
