% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5seurat.R
\name{AppendData}
\alias{AppendData}
\alias{AppendData.character}
\alias{AppendData.H5File}
\title{Append data from an h5Seurat file to a preexisting \code{Seurat} object}
\usage{
AppendData(file, object, ...)

\method{AppendData}{character}(file, object, assays = NULL,
  reductions = NULL, graphs = NULL, overwrite = FALSE,
  verbose = TRUE, ...)

\method{AppendData}{H5File}(file, object, assays = NULL,
  reductions = NULL, graphs = NULL, overwrite = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{file}{Name of h5Seurat or connected h5Seurat file to load}

\item{object}{A \code{Seurat} object to append data to}

\item{...}{Arguments passed to other methods}

\item{assays}{One of:
\itemize{
  \item A character vector with names of assays
  \item A character vector with one or more of \code{counts}, \code{data},
  \code{scale.data} describing which slots of \strong{all assays} to load
  \item A named list where each entry is either the name of an assay or a vector
  describing which slots (described above) to take from which assay
  \item \code{NULL} for all assays
}}

\item{reductions}{One of:
\itemize{
  \item A character vector with names of reductions
  \item \code{NULL} for all reductions
  \item \code{NA} or \code{FALSE} for no reductions
}
\strong{Note}: Only reductions associated with an assay loaded in \code{assays}}

\item{graphs}{One of:
\itemize{
  \item A character vector with names of graphs
  \item \code{NULL} for all graphs
  \item \code{NA} or \code{FALSE} for no reductions
}}

\item{verbose}{Show progress updates}
}
\value{
A \code{Seurat} object with the extra data requested
}
\description{
Append data from an h5Seurat file to a preexisting \code{Seurat} object
}
\seealso{
\code{\link{LoadH5Seurat}} \code{\link{LoadData}}
}
\keyword{internal}
