% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5seurat.R
\name{AssembleObject}
\alias{AssembleObject}
\alias{AssembleAssay}
\alias{AssembleDimReduc}
\alias{AssembleSeuratCommand}
\title{Assemble an object from an h5Seurat file}
\usage{
AssembleAssay(assay, file, cells, slots = c("counts", "data",
  "scale.data"), verbose = TRUE)

AssembleDimReduc(reduction, file, cells, default.assay,
  embeddings = TRUE, loadings = FALSE, projected = FALSE,
  jackstraw = FALSE, verbose = TRUE)

AssembleSeuratCommand(cmd, file, verbose = TRUE)
}
\arguments{
\item{assay, reduction, cmd}{Name of assay, reduction, or command to load}

\item{file}{A connected h5Seurat file to pull the assay from}

\item{cells}{A vector of cell names}

\item{slots}{A vector of slots to load for the assay}

\item{verbose}{Show progress updates}

\item{default.assay}{Name of default assay for this object}

\item{embeddings, loadings, projected, jackstraw}{Pull cell embeddings, feature
loadings, projected loadings, and/or JackStraw information}
}
\value{
\code{AssembleAssay}: An \code{Assay} object

\code{AssembleDimReduc}: A \code{DimReduc} object

\code{AssembleSeuratCommand}: A \code{SeuratCommand} object
}
\description{
Assemble an object from an h5Seurat file
}
\keyword{internal}
