% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{AttachData}
\alias{AttachData}
\title{Attach datasets}
\usage{
AttachData(pkgname = "SeuratData")
}
\arguments{
\item{pkgname}{Name of package}
}
\value{
Invisible \code{NULL}
}
\description{
Attach datasets
}
\keyword{internal}
