% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_data.R
\name{CiteData}
\alias{CiteData}
\title{Get citation information for a dataset}
\usage{
CiteData(ds, lib.loc = NULL)
}
\arguments{
\item{ds}{Name of dataset to get citation information for}

\item{lib.loc}{a character vector with path names of \R libraries, or
    the directory containing the source for \code{package}, or
    \code{NULL}.  The default value of \code{NULL} corresponds to all
    libraries currently known.  If the default is used, the loaded
    packages are searched before the libraries.}
}
\value{
An object of class \code{citation}, inheriting from class
\code{\link[utils]{bibentry}}; see \code{\link[utils]{citation}} for more
details
}
\description{
Get citation information for a dataset
}
\examples{
\dontrun{
CiteData('cbmc')
}

}
\seealso{
\code{\link[utils]{citation}}
}
