% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5seurat.R
\name{GetObject}
\alias{GetObject}
\alias{GetAssays}
\alias{GetDimReducs}
\alias{GetGraphs}
\title{Figure out which objects to load from an h5Seurat file}
\usage{
GetAssays(assays, index)

GetDimReducs(reductions, index, assays = NULL)

GetGraphs(index, assays = NULL)
}
\arguments{
\item{assays}{One of:
\itemize{
  \item A character vector with names of assays
  \item A character vector with one or more of \code{counts}, \code{data},
  \code{scale.data} describing which slots of \strong{all assays} to load
  \item A named list where each entry is either the name of an assay or a vector
  describing which slots (described above) to take from which assay
  \item \code{NULL} for all assays
}}

\item{index}{An h5Seurat index (\code{h5SI}) from \code{\link{IndexH5Seurat}}}

\item{reductions}{A vector of reduction names to load}

\item{graphs}{A vector of graph names to load}
}
\value{
\code{GetAssays}: A named list where each entry is a vector describing
the slots of an assay to load and the names are the assays to load

\code{GetDimReducs}: A vector of reduction names that are derived from
an assay in \code{assays}

\code{GetGraphs}: A vector of graph names that are derived from an
assay in \code{assays}
}
\description{
Figure out which objects to load from an h5Seurat file
}
\seealso{
\code{\link{LoadH5Seurat}} \code{\link{IndexH5Seurat}}
}
\keyword{internal}
