% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{GetPackageData}
\alias{GetPackageData}
\title{Get package datasets}
\usage{
GetPackageData(ds, type = c("memory", "disk"), ...)
}
\arguments{
\item{ds}{Name of a dataset or package to get data for}

\item{type}{Type of data to search for; choose from:
\describe{
\item{memory}{Datasets loaded in memory via \code{\link[utils]{data}}}
\item{disk}{Datasets loaded from disk via \code{\link[SeuratDisk]{LoadH5Seurat}}}
}}

\item{...}{
  Arguments passed on to \code{\link[=NameToPackage]{NameToPackage}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A character vector of all datasets present in \code{ds}
}
\description{
Get package datasets
}
\seealso{
\code{\link[utils]{data}}
}
\keyword{internal}
