% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{IndexObject}
\alias{IndexObject}
\title{Index a saved \code{Seurat} object}
\usage{
IndexObject(directory)
}
\arguments{
\item{directory}{Path to the directory the dataset is stored}
}
\value{
...
}
\description{
Index a saved \code{Seurat} object
}
\seealso{
\code{\link{SaveObject}} \code{\link{LoadObject}}
}
\keyword{internal}
