% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_data.R
\name{InstallData}
\alias{InstallData}
\title{Install a dataset}
\usage{
InstallData(ds, force.reinstall = FALSE, ...)
}
\arguments{
\item{ds}{List of datasets to install}

\item{force.reinstall}{Force reinstall already installed packages}

\item{...}{Extra parameters passed to \code{\link[utils]{install.packages}}}
}
\value{
Invisible \code{NULL}
}
\description{
Install a dataset
}
\examples{
\dontrun{
InstallData('pbmc3k')
}

}
\seealso{
\code{\link{AvailableData}} \code{\link{InstalledData}}
\code{\link{RemoveData}} \code{\link{UpdateData}}
}
