% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_data.R
\name{InstalledData}
\alias{InstalledData}
\title{Get a list of installed datasets}
\usage{
InstalledData()
}
\value{
A dataframe with installed Seurat datasets. Rownames of the dataframe are the actual package names
\describe{
  \item{Dataset}{Name of dataset, usable for other functions in SeuratData (eg. \code{\link{InstallData}})}
  \item{Version}{Version of dataset, generally corresponds to version of Seurat dataset was built with}
  \item{Installed}{Is the dataset installed?}
  \item{InstalledVersion}{Version of dataset installed}
}

Other columns are extra metadata, and may change as new datasets are made available
}
\description{
Get a list of installed datasets
}
\examples{
InstalledData()

}
\seealso{
\code{\link{AvailableData}} \code{\link{InstallData}}
\code{\link{RemoveData}} \code{\link{UpdateData}}
}
