% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_data.R
\name{LoadData}
\alias{LoadData}
\title{Load a dynamic dataset}
\usage{
LoadData(ds, type = c("info", "raw", "processed"), assays = NULL,
  dimreducs = NULL, graphs = NULL)
}
\description{
Load a dynamic dataset
}
\seealso{
\code{\link[base]{data}}
}
