% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5seurat.R
\name{LoadH5Seurat}
\alias{LoadH5Seurat}
\alias{LoadH5Seurat.character}
\alias{LoadH5Seurat.H5File}
\title{Load a saved \code{Seurat} object from an h5Seurat file}
\usage{
LoadH5Seurat(file, ...)

\method{LoadH5Seurat}{character}(file, type = c("info", "object"),
  assays = NULL, reductions = NULL, graphs = NULL, verbose = TRUE,
  ...)

\method{LoadH5Seurat}{H5File}(file, type = c("info", "object"),
  assays = NULL, reductions = NULL, graphs = NULL, verbose = TRUE,
  ...)
}
\arguments{
\item{file}{Name of h5Seurat or connected h5Seurat file to load}

\item{...}{Arguments passed to other methods}

\item{assays}{One of:
\itemize{
  \item A character vector with names of assays
  \item A character vector with one or more of \code{counts}, \code{data},
  \code{scale.data} describing which slots of \strong{all assays} to load
  \item A named list where each entry is either the name of an assay or a vector
  describing which slots (described above) to take from which assay
  \item \code{NULL} for all assays
}}

\item{reductions}{One of:
\itemize{
  \item A character vector with names of reductions
  \item \code{NULL} for all reductions
  \item \code{NA} or \code{FALSE} for no reductions
}
\strong{Note}: Only reductions associated with an assay loaded in \code{assays}}

\item{graphs}{One of:
\itemize{
  \item A character vector with names of graphs
  \item \code{NULL} for all graphs
  \item \code{NA} or \code{FALSE} for no reductions
}}

\item{verbose}{Show progress updates}

\item{meta.data}{Load object metadata?}

\item{commands}{Load command information}

\item{misc}{Load miscellaneous data?}

\item{tools}{Load tool-specific information?}
}
\value{
If \code{type} is info, information about the data contained within the
h5Seurat file. Otherwise, a \code{Seurat} object with the data requested
}
\description{
Load a saved \code{Seurat} object from an h5Seurat file
}
\seealso{
\code{\link{SaveH5Seurat}} \code{\link{IndexH5Seurat}}
}
\keyword{internal}
