% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{LoadObject}
\alias{LoadObject}
\title{Load a saved \code{Seurat} object}
\usage{
LoadObject(directory, type = c("info", "processed"), assays = NULL,
  dimreducs = NULL, graphs = NULL)
}
\arguments{
\item{directory}{Path to the directory the dataset is stored}

\item{type}{How to load the \code{Seurat} object; choose from
\describe{
  \item{info}{Information about the object and what's stored in it}
  \item{raw}{The raw form of the dataset, no other options are evaluated}
  \item{processed}{...}
}}

\item{assays}{...}

\item{dimreducs}{...}

\item{graphs}{...}
}
\value{
...
}
\description{
Load a saved \code{Seurat} object
}
\seealso{
\code{\link{SaveObject}} \code{\link{IndexObject}}
}
\keyword{internal}
