% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_data.R
\name{LoadReference}
\alias{LoadReference}
\title{Load the reference RDS files}
\usage{
LoadReference(path)
}
\arguments{
\item{path}{Pathto the two RDS files}
}
\value{
A list with two entries:
\describe{
 \item{\code{map}}{
  The downsampled reference \code{\link[Seurat]{Seurat}}
  object (for mapping)
 }
 \item{\code{plot}}{The reference \code{Seurat} object (for plotting)}
}
}
\description{
Read in a reference \code{\link[Seurat]{Seurat}} object and annoy index. This
function can read from a file path. In order to read properly,
there must be the following files:
\itemize{
 \item \dQuote{ref.Rds} for the downsampled reference \code{Seurat}
 object (for mapping)
 \item \dQuote{idx.annoy} for the nearest-neighbor index object
}
}
