% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{MakeSpace}
\alias{MakeSpace}
\title{Make a space}
\usage{
MakeSpace(n)
}
\arguments{
\item{n}{Length space should be}
}
\value{
A space (' ') of length \code{n}
}
\description{
Make a space
}
\examples{
\dontrun{
MakeSpace(10)
}

}
\keyword{internal}
