% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{NameToPackage}
\alias{NameToPackage}
\title{Find dataset packages based on name}
\usage{
NameToPackage(ds)
}
\arguments{
\item{ds}{Names of datasets}
}
\value{
A vector of package names based on dataset names
}
\description{
Find dataset packages based on name
}
\examples{
\dontrun{
NameToPackage('cbmc')
NameToPackage('pbmc3k.SeuratData')
NameToPackage('notadataset')
}

}
\keyword{internal}
