% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5seurat.R
\name{print.h5SI}
\alias{print.h5SI}
\alias{print}
\title{Print index information for an h5Seurat Index (h5SI)}
\usage{
\method{print}{h5SI}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{x}
}
\description{
Print index information for an h5Seurat Index (h5SI)
}
\keyword{internal}
