% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_data.R
\name{RemoveData}
\alias{RemoveData}
\title{Remove a dataset}
\usage{
RemoveData(pkgs, lib)
}
\arguments{
\item{pkgs}{List of dataset to remove}

\item{...}{Arguments passed on to \code{utils::remove.packages}
\describe{
  \item{pkgs}{a character vector with the names of the packages to be removed.}
  \item{lib}{a character vector giving the library directories to remove the
    packages from.  If missing, defaults to the first element in
    \code{\link{.libPaths}()}.}
}}
}
\description{
Remove a dataset
}
