% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5seurat.R
\name{SaveH5Seurat}
\alias{SaveH5Seurat}
\alias{SaveH5Seurat.default}
\alias{SaveH5Seurat.Seurat}
\title{Save a \code{Seurat} object to a h5Seurat file}
\usage{
SaveH5Seurat(object, filename, overwrite = FALSE, verbose = TRUE, ...)

\method{SaveH5Seurat}{default}(object, filename, overwrite = FALSE,
  verbose = TRUE, ...)

\method{SaveH5Seurat}{Seurat}(object, filename = paste0(Project(object =
  object), ".h5Seurat"), overwrite = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{filename}{Name for h5Seurat file}

\item{overwrite}{If an h5Seurat file of name \code{filename} exists, overwrite it?}

\item{verbose}{Show progress updates}

\item{...}{Arguments passed to other methods}
}
\value{
Invisibly returns the name of the h5Seurat file
}
\description{
Save a \code{Seurat} object to a h5Seurat file
}
\seealso{
\code{\link{LoadH5Seurat}} \code{\link{IndexH5Seurat}}
}
\keyword{internal}
