% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{SaveObject}
\alias{SaveObject}
\title{Save a \code{Seurat} object into component Rds files}
\usage{
SaveObject(object, directory = file.path(getwd(), Project(object =
  object)), verbose = TRUE)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{directory}{Directory to save the object to}

\item{verbose}{Show progress messages}
}
\value{
Invisibly return the directory that the \code{Seurat} object was stored in
}
\description{
Save a \code{Seurat} object into component Rds files
}
\seealso{
\code{\link{IndexObject}} \code{\link{LoadObject}}
}
\keyword{internal}
