% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{SaveSeuratObject}
\alias{SaveSeuratObject}
\title{Save a Seurat object into component Rds files}
\usage{
SaveSeuratObject(object, directory = file.path(getwd(), Project(object =
  object)), verbose = TRUE)
}
\arguments{
\item{...}{...}
}
\value{
Invisibly return the directory that the Seurat object was stored in
}
\description{
Save a Seurat object into component Rds files
}
\keyword{internal}
