% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set-if-null}
\alias{set-if-null}
\alias{\%||\%}
\title{Set a default value if an object is null}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{An object to set if it's null}

\item{rhs}{The value to provide if x is null}
}
\value{
rhs if lhs is null, else lhs
}
\description{
Set a default value if an object is null
}
\examples{
\dontrun{
4 \%||\% 5
NULL \%|| 5
}

}
\references{
\url{https://adv-r.hadley.nz/functions.html#missing-arguments}
}
\author{
Hadley Wickham
}
\keyword{internal}
