% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SeuratData-package}
\alias{SeuratData}
\alias{SeuratData-package}
\title{Install and manage Seurat datasets}
\description{
Single cell RNA sequencing datasets can be large, consisting of matrices
    that contain expression data for several thousand features across several thousand
    cells. This package is designed to easily install, manage, and learn about various single-cell
    datasets, provided Seurat objects and distributed as independent packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.satijalab.org/seurat}
  \item \url{https://github.com/satijalab/seurat-data}
  \item Report bugs at \url{https://github.com/satijalab/seurat-data/issues}
}

}
\author{
\strong{Maintainer}: Paul Hoffman \email{phoffman@nygenome.org} (0000-0002-7693-8957)

Authors:
\itemize{
  \item Rahul Satija \email{rsatija@nygenome.org} (0000-0001-9448-8833)
}

}
\keyword{internal}
