% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_data.R
\name{UpdateData}
\alias{UpdateData}
\title{Update datasets}
\usage{
UpdateData(ask = TRUE, lib.loc = NULL)
}
\arguments{
\item{ask}{logical indicating whether to ask the user to select
    packages before they are downloaded and installed, or the character
    string \code{"graphics"}, which brings up a widget to allow the user
    to (de-)select from the list of packages which could be updated.
    (The latter value only works on systems with a GUI version of
    \code{\link{select.list}}, and is otherwise equivalent to
    \code{ask = TRUE}.)
  }

\item{lib.loc}{character vector describing the location of R
    library trees to search through (and update packages therein), or
    \code{NULL} for all known trees (see \code{\link{.libPaths}}).}
}
\value{
Invisible \code{NULL}
}
\description{
Update datasets
}
\seealso{
\code{\link{AvailableData}} \code{\link{InstallData}} \code{\link{InstalledData}} \code{\link{RemoveData}}
}
