% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{UpdateManifest}
\alias{UpdateManifest}
\title{Update the available package manifest}
\usage{
UpdateManifest()
}
\value{
Nothing, updates the manifest in the package environment
}
\description{
Update the available package manifest
}
\keyword{internal}
