% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5seurat.R
\name{ValidateH5SI}
\alias{ValidateH5SI}
\title{Validate an h5Seurat Index}
\usage{
ValidateH5SI(x)
}
\arguments{
\item{x}{An h5Seurat Index (h5SI)}
}
\value{
\code{x} if valid, otherwise a modified \code{x} to be valid
}
\description{
Validate an h5Seurat Index
}
\keyword{internal}
