% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5seurat.R
\name{WriteH5List}
\alias{WriteH5List}
\title{Write a list as a series of HDF5 groups and datasets}
\usage{
WriteH5List(x, name, hgroup)
}
\arguments{
\item{x}{A list}

\item{name}{Name to save dataset as}

\item{hgroup}{An HDF5 file or group (\code{H5File} or
\code{H5Group} objects from hdf5r)}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Write a list as a series of HDF5 groups and datasets
}
\keyword{internal}
