% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5seurat.R
\name{as.list.H5Group}
\alias{as.list.H5Group}
\alias{as.list}
\title{Read an HDF5 group into a ist}
\usage{
\method{as.list}{H5Group}(x, ...)
}
\arguments{
\item{x}{An \code{H5Group} object (from hdf5r)}

\item{...}{Ignored}
}
\value{
A list with the data contained within \code{x}; if the HDF5 attribute
\code{s4class} is set and is a class, will return an object of class \code{s4class}
instead
}
\description{
Read an HDF5 group into a ist
}
\keyword{internal}
