% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_data.R
\name{AvailableData}
\alias{AvailableData}
\title{Get list of available datasets}
\usage{
AvailableData()
}
\value{
A dataframe with available Seurat datasets. Rownames of the dataframe are the actual package names
\describe{
  \item{Dataset}{Name of dataset, usable for other functions in SeuratData (eg. \code{\link{InstallData}})}
  \item{Version}{Version of dataset, generally corresponds to version of Seurat dataset was built with}
  \item{Installed}{Is the dataset installed?}
  \item{InstalledVersion}{Version of dataset installed, \code{NA} if not installed}
}

Other columns are extra metadata, and may change as new datasets are made available
}
\description{
Get list of available datasets
}
\examples{
AvailableData()

}
\seealso{
\code{\link{InstallData}} \code{\link{InstalledData}}
\code{\link{RemoveData}} \code{\link{UpdateData}}
}
