% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5seurat.R
\name{IndexH5Seurat}
\alias{IndexH5Seurat}
\alias{IndexH5Seurat.character}
\alias{IndexH5Seurat.H5File}
\title{Index an H5Seurat object}
\usage{
IndexH5Seurat(file, ...)

\method{IndexH5Seurat}{character}(file, ...)

\method{IndexH5Seurat}{H5File}(file, ...)
}
\arguments{
\item{...}{...}
}
\value{
...
}
\description{
Index an H5Seurat object
}
\seealso{
\code{\link{LoadH5Seurat}} \code{\link{SaveH5Seurat}}
}
\keyword{internal}
