% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_data.R
\name{LoadData}
\alias{LoadData}
\title{Modularly load a dataset}
\usage{
LoadData(ds, type = c("info", "raw", "processed"), assays = NULL,
  reductions = NULL, graphs = NULL, verbose = TRUE)
}
\arguments{
\item{ds}{Optional name of dataset to load}

\item{type}{How to load the \code{Seurat} object; choose from
\describe{
  \item{info}{Information about the object and what's stored in it}
  \item{raw}{The raw form of the dataset, no other options are evaluated}
  \item{processed}{The proccessed data, modular loading avaible by setting other parameters}
}}
}
\description{
Modularly load a dataset
}
\seealso{
\code{\link[utils]{data}}
}
