% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_data.R
\name{LoadData}
\alias{LoadData}
\title{Modularly load a dataset}
\usage{
LoadData(ds, type = "info", assays = NULL, reductions = NULL,
  graphs = NULL, verbose = TRUE)
}
\arguments{
\item{ds}{Name of dataset to load}

\item{type}{How to load the \code{Seurat} object; choose from
\describe{
  \item{info}{
  Information about the object and what's stored in it; to specify a dataset
  to get information for, pass 'info_\code{name of dataset}' based on the name
  of the dataset in the \code{other.datasets} column of the data manifest
  }
  \item{raw}{The raw form of the dataset, no other options are evaluated}
  \item{...}{
  Name of dataset in the \code{other.datasets} column of the data manifest
  corresponding to the dataset desired (eg. 'processed' for ifnb)
  }
}}

\item{assays}{One of:
\itemize{
  \item A character vector with names of assays
  \item A character vector with one or more of \code{counts}, \code{data},
  \code{scale.data} describing which slots of \strong{all assays} to load
  \item A named list where each entry is either the name of an assay or a vector
  describing which slots (described above) to take from which assay
  \item \code{NULL} for all assays
}}

\item{reductions}{One of:
\itemize{
  \item A character vector with names of reductions
  \item \code{NULL} for all reductions
  \item \code{NA} or \code{FALSE} for no reductions
}
\strong{Note}: Only reductions associated with an assay loaded in \code{assays}}

\item{graphs}{One of:
\itemize{
  \item A character vector with names of graphs
  \item \code{NULL} for all graphs
  \item \code{NA} or \code{FALSE} for no reductions
}}

\item{verbose}{Show progress updates}
}
\value{
If \code{type} is info, information about the data contained within the
h5Seurat file. Otherwise, a \code{Seurat} object with the data requested
}
\description{
Modularly load a dataset
}
\examples{
\dontrun{
LoadData('ifnb')
ifnb.raw <- LoadData('ifnb', type = 'raw')
ifnb.proccessed <- LoadData('ifnb', type = 'processed')
}

}
\seealso{
\code{\link[utils]{data}}
}
