% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_data.R
\name{LoadData}
\alias{LoadData}
\title{Modularly load a dataset}
\usage{
LoadData(ds, type = c("info", "raw", "processed"), assays = NULL,
  reductions = NULL, graphs = NULL, verbose = TRUE)
}
\arguments{
\item{ds}{Optional name of dataset to load}

\item{type}{How to load the \code{Seurat} object; choose from
\describe{
  \item{info}{Information about the object and what's stored in it}
  \item{raw}{The raw form of the dataset, no other options are evaluated}
  \item{processed}{The proccessed data, modular loading avaible by setting other parameters}
}}

\item{assays}{One of:
\itemize{
  \item A character vector with names of assays
  \item A character vector with one or more of \code{counts}, \code{data},
  \code{scale.data} describing which slots of the \strong{default assay}
  \item A named list where each entry is either the name of an assay or a vector
  describing which slots (described above) to take from which assay
  \item \code{NULL} for all assays
}}

\item{reductions}{One of:
\itemize{
  \item A character vector with names of reductions
  \item \code{NULL} for all reductions
  \item \code{NA} or \code{FALSE} for no reductions
}
\strong{Note}: Only reductions associated with an assay loaded
in \code{assays}}

\item{graphs}{One of:
\itemize{
  \item A character vector with names of graphs
  \item \code{NULL} for all graphs
  \item \code{NA} or \code{FALSE} for no reductions
}}

\item{verbose}{Show progress updates}
}
\value{
If \code{type} is info, information about the data contained within the
h5Seurat file. Otherwise, a \code{Seurat} object with the data requested
}
\description{
Modularly load a dataset
}
\seealso{
\code{\link[utils]{data}}
}
