% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_data.R
\name{LoadData}
\alias{LoadData}
\title{Modularly load a dataset}
\usage{
LoadData(ds, type = "default")
}
\arguments{
\item{ds}{Optional name of dataset to load}

\item{type}{How to load the \code{Seurat} object; choose from either
'default' for the default dataset or any dataset listed in the
\code{other.datasets} section of the data manifest}
}
\value{
A \code{Seurat} object with the dataset asked for
}
\description{
Modularly load a dataset
}
\seealso{
\code{\link[utils]{data}}
}
