% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_data.R
\name{RemoveData}
\alias{RemoveData}
\title{Remove a dataset}
\usage{
RemoveData(ds, lib)
}
\arguments{
\item{ds}{List of datasets to remove}

\item{lib}{a character vector giving the library directories to remove the
    packages from.  If missing, defaults to the first element in
    \code{\link{.libPaths}()}.}
}
\description{
Remove a dataset
}
\examples{
\dontrun{
RemoveData('pbmc3k')
}

}
\seealso{
\code{\link{AvailableData}} \code{\link{InstallData}}
\code{\link{InstalledData}} \code{\link{UpdateData}}
}
