% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SeuratData-package}
\alias{SeuratData}
\alias{SeuratData-package}
\title{Install and manage Seurat datasets}
\description{
Single cell RNA sequencing datasets can be large, consisting of matrices that contain expression data for several thousand features across several thousand cells. This package is designed to easily install, manage, and learn about various single-cell datasets, provided Seurat objects and distributed as independent packages.
}
\section{Package options}{


SeuratData uses the following options to control behaviour, users can configure
these with \code{\link[base]{options}}:

\itemize{
  \item `SeuratData.repo.use`: Set the location where the SeuratData datasets
  are stored. Users generally should not modify.
  \item `SeuratData.manifest.cache`: Cache the data manifest whenever we talk
  to the data repository; note, setting to \code{FALSE} will simply prevent
  SeuratData from caching the manifest, not from reading a previously cached
  manifest
  \item `SeuratData.roaming`: For Windows users, use a roaming profile directory
  for domain users. See \url{https://en.wikipedia.org/wiki/Roaming_user_profile}
  for a brief overview and Microsoft's documentation for greater detail
}
}

\seealso{
Useful links:
\itemize{
  \item \url{http://www.satijalab.org/seurat}
  \item \url{https://github.com/satijalab/seurat-data}
  \item Report bugs at \url{https://github.com/satijalab/seurat-data/issues}
}

}
\author{
\strong{Maintainer}: Paul Hoffman \email{phoffman@nygenome.org} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID})

Authors:
\itemize{
  \item Rahul Satija \email{rsatija@nygenome.org} (\href{https://orcid.org/0000-0001-9448-8833}{ORCID})
  \item Andrew Butler \email{abutler@nygenome.org} (\href{https://orcid.org/0000-0003-3608-0463}{ORCID})
}

Other contributors:
\itemize{
  \item Austin Hartman \email{ahartman@nygenome.org} (\href{https://orcid.org/0000-0001-7278-1852}{ORCID}) [contributor]
  \item Yuhan Hao \email{yhao@nygenome.org} (\href{https://orcid.org/0000-0002-1810-0822}{ORCID}) [contributor]
  \item Gesmira Molla \email{gmolla@nygenome.org} (\href{https://orcid.org/0000-0002-8628-5056}{ORCID}) [contributor]
}

}
\keyword{internal}
